+++ backend/src/utils/email-validator-adapter.ts
